+++ test/runtests.jl
@@ -15,6 +15,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_checkpoint.jl")
         include("unit/test_sqw_core.jl")
         include("unit/test_sqw_gates.jl")
+        include("unit/test_gpu_llg.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")